feat: Add data fetching and preliminary plotting
To enable the user to fetch and save the USA's SSA first name data
locally. The data is saved loaded into a pandas dataframe and then saved
as a parquet in the data/ folder.

Plotting is preliminary but accomplishes the main goal of visualizing a
name or a list of names popularity throughout time.