feat: Add random name generator
Often it is hard to find names one likes, so you can start randomly
generating names to start looking for names you might like. Has
capabilities to slice the data to narrow search down to desirable
features, such as gender of name, first letter of name, popularity of
name (rank), and years name was in vogue.

Also cleaned up files by adding a utils.py for general purpose
functions. The plots.py was improved so that the user can now provide
either a single string or list of len 1 for the genders, which is then
extrapolated to match the length of names given.